scraper/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
aiosqlite==0.21.0
alpha_vantage==3.0.0
annotated-types==0.7.0
anyio==4.9.0
astroid==3.3.9
asttokens==3.0.0
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.3
bintrees==2.2.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colormath==3.0.0
contourpy==1.3.1
CoolProp==6.7.0
cryptography==43.0.3
cycler==0.12.1
dataclasses-json==0.6.7
DateTime==5.5
decorator==5.2.1
dill==0.3.9
distro==1.9.0
dydantic==0.0.8
emoji==2.14.1
executing==2.2.0
faiss-cpu==1.10.0
filelock==3.18.0
flexcache==0.3
flexparser==0.4
fonttools==4.56.0
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2025.3.0
fuzzywuzzy==0.18.0
geographiclib==2.0
geopy==2.4.1
google==3.0.0
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-vision==3.10.1
google_search_results==2.4.2
googleapis-common-protos==1.69.2
googlemaps==4.10.0
googlesearch-python==1.3.0
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
gspread==6.2.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
inflection==0.5.1
ipython_pygments_lexers==1.1.1
isort==6.0.1
jedi==0.19.2
jellyfish==1.1.3
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jsonschema_rs==0.29.1
kiwisolver==1.4.8
langchain==0.3.21
langchain-community==0.3.20
langchain-core==0.3.46
langchain-experimental==0.3.4
langchain-openai==0.3.9
langchain-text-splitters==0.3.7
langgraph==0.3.18
langgraph-checkpoint==2.0.21
langgraph-checkpoint-sqlite==2.0.6
langgraph-cli==0.1.77
langgraph-prebuilt==0.1.3
langgraph-sdk==0.1.58
langsmith==0.3.18
language_tool_python==2.9.0
Levenshtein==0.27.1
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mccabe==0.7.0
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.2.0
multitasking==0.0.11
mypy-extensions==1.0.0
names==0.3.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauth2client==4.1.3
oauthlib==3.2.2
openai==1.72.0
opencv-python==4.11.0.86
orjson==3.10.15
packaging==24.2
pandas==2.2.3
parso==0.8.4
peewee==3.17.9
pexpect==4.9.0
pillow==11.1.0
Pint==0.24.4
platformdirs==4.3.7
prompt_toolkit==3.0.50
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.4
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
pylint==3.3.6
pyparsing==3.2.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.27.1
pytz==2025.1
PyYAML==6.0.2
Quandl==3.7.0
RapidFuzz==3.12.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rpds-py==0.24.0
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.2
serpapi==0.1.5
simpy==4.1.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.39
sse-starlette==2.1.3
stack-data==0.6.3
starlette==0.46.1
structlog==25.2.0
sympy==1.13.1
tavily-python==0.5.1
tenacity==9.0.0
termcolor==2.5.0
textblob==0.19.0
threadpoolctl==3.6.0
tiktoken==0.9.0
toml==0.10.2
tomlkit==0.13.2
torch==2.6.0
tqdm==4.67.1
traitlets==5.14.3
triton==3.2.0
trustcall==0.0.38
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
validators==0.34.0
watchfiles==1.0.4
wcwidth==0.2.13
wikipedia==1.4.0
word2number==1.1
xmltodict==0.14.2
yarl==1.18.3
yfinance==0.2.54
zipfile36==0.1.3
zope.interface==7.2
zstandard==0.23.0